централизованная обработка ошибок

+++ src/controllers/cards.ts
@@ -1,55 +1,44 @@
-import { Request, Response } from 'express';
+import { Request, Response, NextFunction } from 'express';
 import Card from '../models/cards';
 import NotFoundError from '../errors/not_found_error';
-import
-{
-  BAD_REQUEST_ERROR_STATUS_CODE,
-  NOT_FOUND_ERROR_STATUS_CODE,
-  INTERNAL_SERVER_ERROR_STATUS_CODE,
-  FORBIDDEN_ERROR_STATUS_CODE,
-} from '../constants';
+import BadRequestError from '../errors/bad_request_error';
+import ForbiddenError from '../errors/forbidden_error';
 
-export const getCards = (req: Request, res: Response) => Card.find({})
+export const getCards = (req: Request, res: Response, next: NextFunction) => Card.find({})
   .then((cards) => res.send({ data: cards }))
-  .catch(() => res.status(INTERNAL_SERVER_ERROR_STATUS_CODE).send({ message: 'На сервере произошла ошибка' }));
+  .catch(next);
 
-export const createCard = (req: Request, res: Response) => {
+export const createCard = (req: Request, res: Response, next: NextFunction) => {
   const { name, link } = req.body;
   return Card.create({ name, link, owner: req.user._id })
-    .then((card) => {
-      if (!card) {
-        throw new NotFoundError('Карточка не найдена');
-      }
-      return res.send({ data: card });
-    })
+    .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST_ERROR_STATUS_CODE).send({ message: 'Переданы некорректные данные при создании карточки' });
+        next(new BadRequestError('Переданы некорректные данные при создании карточки'));
       }
-      return res.status(INTERNAL_SERVER_ERROR_STATUS_CODE).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
 
-export const deleteCardById = (req: Request, res: Response) => {
+export const deleteCardById = (req: Request, res: Response, next: NextFunction) => {
   Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
-        throw new NotFoundError('Карточка не найдена');
+        next(new NotFoundError('Карточка не найдена'));
       }
-      if (card.owner.toString() === req.user._id.toString()) {
+      if (card && card.owner.toString() === req.user._id.toString()) {
         card.delete();
-        return res.send({ message: 'Карточка удалена' });
-      } return res.status(FORBIDDEN_ERROR_STATUS_CODE).send({ message: 'Недостаточно прав для удаления карточки' });
+      } next(new ForbiddenError('Недостаточно прав для удаления карточки'));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(NOT_FOUND_ERROR_STATUS_CODE).send({ message: 'Карточка с указанным _id не найдена' });
+        next(new NotFoundError('Карточка с указанным _id не найденаа'));
       }
-      return res.status(INTERNAL_SERVER_ERROR_STATUS_CODE).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
 
-export const likeCard = (req: Request, res: Response) => {
+export const likeCard = (req: Request, res: Response, next: NextFunction) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -59,16 +48,16 @@ export const likeCard = (req: Request, res: Response) => {
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST_ERROR_STATUS_CODE).send({ message: 'Переданы некорректные данные для проставления лайка' });
+        next(new BadRequestError('Переданы некорректные данные для проставления лайка'));
       }
       if (err.name === 'CastError') {
-        res.status(NOT_FOUND_ERROR_STATUS_CODE).send({ message: 'Передан несуществующий _id карточки' });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       }
-      return res.status(INTERNAL_SERVER_ERROR_STATUS_CODE).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
 
-export const dislikeCard = (req: Request, res: Response) => {
+export const dislikeCard = (req: Request, res: Response, next: NextFunction) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -78,11 +67,11 @@ export const dislikeCard = (req: Request, res: Response) => {
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST_ERROR_STATUS_CODE).send({ message: 'Переданы некорректные данные для удаления лайка' });
+        next(new BadRequestError('Переданы некорректные данные для удаления лайка'));
       }
       if (err.name === 'CastError') {
-        res.status(NOT_FOUND_ERROR_STATUS_CODE).send({ message: 'Передан несуществующий _id карточки' });
+        next(new NotFoundError('Передан несуществующий _id карточки'));
       }
-      return res.status(INTERNAL_SERVER_ERROR_STATUS_CODE).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
